[FIX] Added dependency injection to the Task class. Added tests for the BuildFolderClean task.
